database_session_demo/requirements.txt
google-adk
psycopg2-binary
python-dotenv
